backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
opencv-python-headless==4.9.0.80
numpy==1.26.4
httpx==0.27.0
pydantic==2.7.1
google-generativeai==0.7.2
python-dotenv==1.0.1
